+++ generators/app/templates/benefit/app/common/base_service.js
@@ -1,0 +1,7 @@
+class BaseService {
+  constructor () {
+    this.ctx = require('../../server').context
+  }
+}
+
+module.exports = BaseService