+++ generators/app/templates/benefit/server.js
@@ -9,7 +9,6 @@ const views = require('koa-views')
 const path = require('path')
 const config = require('./config')
 const appModule = require('./app')
-const dir = require('dir_filenames')
 
 const app = new Koa()
 
@@ -17,8 +16,6 @@ const keys = Object.keys(appModule.context)
 keys.map(key => {
   Object.defineProperty(app.context, key, appModule.context[key])
 })
-const services = dir(`${config.appRoot}/app/services`)
-const models = dir(`${config.appRoot}/app/models`)
 
 app.use(async (ctx, next) => {
   if (ctx.request.method === 'OPTIONS') {
@@ -29,16 +26,8 @@ app.use(async (ctx, next) => {
   ctx.set('Access-Control-Max-Age', 86400000)
   ctx.set('Access-Control-Allow-Methods', 'OPTIONS, GET, PUT, POST, DELETE')
   ctx.set('Access-Control-Allow-Headers', 'x-requested-with, accept, origin, content-type')
-  ctx.service = {}
-  services.map(file => {
-    let name = file.split('/').pop().replace(/\.\w+$/, '')
-    ctx.service[name] = require(file)
-  })
-  ctx.model = {}
-  models.map(file => {
-    let name = file.split('/').pop().replace(/\.\w+$/, '')
-    ctx.model[name] = require(file)
-  })
+  ctx.service = appModule.service
+  ctx.model = appModule.model
   try {
     await next()
   } catch (err) {